feat(install): expose sidecar dispatch contract
