+++ llms.txt
@@ -690,7 +690,7 @@ tools:
 ### Adapter Install Lifecycle (`dcc_mcp_core.install_lifecycle`)
 
 - Import-light helpers for adapter installers/uninstallers that must not load `dcc_mcp_core._core` from an install root they may remove.
-- `query_runtime_state(registry_dir=None, dcc_type=None, role=None, install_root=None, include_dead=True)` reads `services.json` with stdlib only; sidecar rows expose top-level `dispatch_status`, `dispatch_ready`, `host_rpc_uri`, `host_rpc_scheme`, `failure_stage`, and `failure_reason`. Gateway Admin mirrors the same fields on `GET /admin/api/workers`; `gateway://instances`, `GET /v1/instances`, and `/admin/api/instances` expose a nested `dispatch` object with `reported`, `status`, `ready`, host-RPC metadata, and failure metadata.
+- `query_runtime_state(registry_dir=None, dcc_type=None, role=None, install_root=None, include_dead=True)` reads `services.json` with stdlib only; sidecar rows expose top-level `dispatch_status`, `dispatch_ready`, `host_rpc_uri`, `host_rpc_scheme`, `failure_stage`, and `failure_reason` for compatibility and a nested `dispatch` object (`reported`, `status`, `ready`, `ready_at_unix`, `host_rpc_uri`, `host_rpc_scheme`, `failure_stage`, `failure_reason`) for new callers. Gateway Admin mirrors the worker fields on `GET /admin/api/workers`; `gateway://instances`, `GET /v1/instances`, and `/admin/api/instances` expose a nested `dispatch` object with the same readiness, host-RPC, and failure metadata.
 - `sidecar_readiness_status(...)` returns a one-shot sidecar dispatch verdict (`ready`, `missing`, `booting`, `unavailable`, or `dead`); `wait_for_sidecar_ready(...)` polls the same verdict for bounded installer/supervisor checks without importing native core.
 - Pass `probe_tool="maya_diagnostics__ping"` (or CLI `--probe-tool`) to require one real `tools/call` through the sidecar before reporting ready; optional `probe_arguments` / `--probe-args-json` must be a JSON object. Probe failures keep polling until timeout in `wait_for_sidecar_ready(...)`.
 - `stop_runtime_entries(..., role="per-dcc-sidecar", include_host_processes=False)` requests sidecar stop without terminating the owning DCC host by default.
